Extract applyQueryParamsToUserSettings
